Starting to write my own OpenGL renderer

+++ Physics2D/src/p2Body.h
@@ -73,7 +73,7 @@ public:
 	inline p2BodyType GetType() const { return m_type; }
 	inline void SetVelocity(p2Vec2 vel) { if (m_type == p2_kinematicBody)return; m_linearVelocity = vel; }
 	inline void SetPosition(p2Vec2 newPos) { m_position = newPos; }
-	inline int GetFixtureCount() { return m_fixtures.size(); }
+	inline size_t GetFixtureCount() { return m_fixtures.size(); }
 	inline p2Fixture* GetFixture(int index) { return m_fixtures[index]; }
 
 	//TODO: Make this better
