+++ Physics2D/src/p2CircleShape.cpp
@@ -12,7 +12,7 @@ p2CircleShape::p2CircleShape()
 p2MassData* p2CircleShape::ComputeMass(float density)
 {
 	m_massData = new p2MassData();
-	m_massData->mass = density * M_PI * m_radius * m_radius;
+	m_massData->mass = density * (float)M_PI * m_radius * m_radius;
 	m_massData->centre = m_pos;
 	m_massData->I = m_massData->mass * (0.5f * m_radius * m_radius + p2Dot(m_pos, m_pos));
 	return m_massData;
